Add files via upload
The new thing about this is that it has an entry box, the 3d polygons can spin correctly (I tested it out with a square and it seemed to work fine) and they can rotate about some axis. A new object called shape3d is also created which takes in a list of poly3d objects and uses that to essentially model and 3d shape. shape3d can display, erase, update, move, spin and rotate although the spinning function is a bit weird.