Add files via upload
The new thing about this version is that it has entries and textboxes with scrollbars in it (the scrollbar for the entries is kind of dodgy). A fps and timer function have been implemented in the window class, allowing the user to better manage time with the app they are creating. The new things I want to implement are menus, scrollbars for the entire window, allowing users to implement audio, and mouse click triggers.