Add files via upload
The difference about this one is that all of the widgets and shapes have a displayed attribute to keep track of whether they are displayed on the screen. The other big difference is for the shape3d object. Now when you spin it the layering of the polygons is not so much of a problem, though it is still a bit glitchy.